extract obj2brain into opengever.testing.
