+++ opengever/document/tests/test_indexers.py
@@ -1,6 +1,7 @@
 from Products.CMFCore.utils import getToolByName
 from opengever.document.checkout.manager import CHECKIN_CHECKOUT_ANNOTATIONS_KEY
 from opengever.document.testing import OPENGEVER_DOCUMENT_FUNCTIONAL_TESTING
+from opengever.testing import obj2brain
 from plone.app.testing import TEST_USER_ID
 from plone.app.testing import setRoles
 from plone.dexterity.utils import createContentInContainer
@@ -12,17 +13,6 @@ import unittest2 as unittest
 def getindexDataForObj(obj):
     catalog = getToolByName(obj, 'portal_catalog')
     return catalog.getIndexDataForRID(obj2brain(obj).getRID())
-
-
-def obj2brain(obj):
-    catalog = getToolByName(obj, 'portal_catalog')
-    query = {'path': {'query': '/'.join(obj.getPhysicalPath()),
-                  'depth': 0}}
-    brains = catalog(query)
-    if len(brains) == 0:
-        raise Exception('Not in catalog: %s' % obj)
-    else:
-        return brains[0]
 
 
 class TestDocumentIntegration(unittest.TestCase):
