+++ opengever/testing/__init__.py
@@ -9,6 +9,7 @@ except pkg_resources.DistributionNotFound:
 
 else:
     from opengever.testing.helpers import create_plone_user
+    from opengever.testing.helpers import obj2brain
     from opengever.testing.sql import create_client
     from opengever.testing.sql import create_ogds_user
     from opengever.testing.sql import set_current_client_id
